requirements.txt
Flask
Werkzeug
Pillow
PyPDF2
gunicorn